Fixed game restart
