+++ src/main/java/classes/menus/gamePanels/SinglePlayerPanel.java
@@ -16,7 +16,7 @@ import java.util.List;
 
 public class SinglePlayerPanel extends GamePanel {
     private final Snake snake = new Snake(
-            Direction.RIGHT,
+            Direction.LEFT,
             GamePanel.GAME_PANEL_WIDTH / 2 / Body.BODY_SIZE * Body.BODY_SIZE,
             GamePanel.GAME_PANEL_HEIGHT / 2 / Body.BODY_SIZE * Body.BODY_SIZE);
